+++ noodles-bam/CHANGELOG.md
@@ -24,6 +24,8 @@
 
   * bam/record: Replace `QualityScores` with `sam::record::QualityScores`.
 
+  * bam/record: Replace `Sequence` with `sam::record::Sequence`.
+
   * bam/reader/record: Clear quality scores if all scores are missing.
 
     This previously filled the quality scores with scores of 255, but this just
@@ -51,6 +53,10 @@
   * bam/record: Remove `QualityScores`.
 
     Use `sam::record::QualityScores` instead.
+
+  * bam/record: Remove `Sequence`.
+
+    Use `sam::record::Sequence` instead.
 
 ## 0.16.0 - 2022-03-02